update read me and rename file.
