+++ README.md
@@ -23,14 +23,14 @@ If you have selected install_dependencies during the prompt, you can skip this p
         ```
         npm run build
         ```
-    2. Run and modify the `usage.py` sample dash app:
+    2. Run and modify the `example.py` sample dash app:
         ```
-        python usage.py
+        python example.py
         ```
     3. Visit http://localhost:8050 in your web browser
 
 - Write tests for your component.
-    - A sample test is available in `tests/test_usage.py`, it will load `usage.py` and you can then automate interactions with selenium.
+    - A sample test is available in `tests/test_example.py`, it will load `example.py` and you can then automate interactions with selenium.
     - Run the tests with `pytest tests`.
     - The Dash team uses these types of integration tests extensively. Browse the Dash component code on GitHub for more examples of testing (e.g. https://github.com/plotly/dash-core-components)
 - Add custom styles to your component by putting your custom CSS files into your distribution folder (`uk_gov_dash_components`).
@@ -52,7 +52,8 @@ If you have selected install_dependencies during the prompt, you can skip this p
     See [PyPA](https://packaging.python.org/guides/distributing-packages-using-setuptools/#packaging-your-project)
     for more information.
 
-3. Test your tarball by copying it into a new environment and installing it locally:
+3. Test your tarball by copying it's full file path and installing it locally into a new environment by running:
     ```
-    pip install uk_gov_dash_components-1.0.0-beta.7.tar.gz
+    pip install <full file path>
     ```
+    eg. `pip install D:\Users\user\Desktop\Source Code Store\gov-uk-dash-react-components\dist\uk_gov_dash_components-1.2.1.tar.gz`
